Create new file for the gatway functions.
