+++ crates/papyrus_load_test/src/lib.rs
@@ -1,0 +1,1 @@
+pub mod gateway_functions;
